[MIRROR] Add Show Hiddenprints action to VV on atoms (#5212)
* Add Show Hiddenprints action to VV on atoms (#58657)

This displays the hiddenprint log from /datum/component/forensics in a simple to view form, sorted by last touch. I didn't make it fancy, it's very bare bones.

I could have refactored hiddenprints to use lists in lists, which would make getting data out a lot easier, but I worry about the additional memory overhead on every atom that gets touched, so I left it as strings in lists. I did reformat it a little bit, however, to make it simpler to sort, and it looks a little nicer. Namely moving the timestamp in front of the name.

It can help a lot in admin investigations on whodunnit. Right now they have to look at the forensics component, and that's if they even know about it. This way it is discoverable by using the VV tool.

* Add Show Hiddenprints action to VV on atoms

Co-authored-by: Jonathan Rubenstein <jrubcop@gmail.com>