[MIRROR] Fixes the color consistency of swat helmets. (#5209)
* fixes the color consistency of swat helmets (#58648)

This PR adjusts the SWAT helmet to have its colors be consistent between its front, side and back sprites.

Someone pointed out to me that the swat helmet looked wrong on mobs for the side and back sprites and yeah, it does.

Bottom's the tweaked version. The back is also tweaked to have the same colors. I chose the black of the front over the gray of the side/back because it distinguishes the helmet from a regular sec helmet.

* Fixes the color consistency of swat helmets.

Co-authored-by: CRITAWAKETS <sebastienracicot@hotmail.com>