[MIRROR] Stops ethereal from crystalizing while eaten  (#5210)
* Stops ethereal from crystalizing while eaten (#58650)

Checks if an ethereal is in an item or mob before making the crystal. Stops weird cases where one might just appear outside a space dragon or create legions that dropped no corpses. fixes #58643

Look if someone wants to code the process of an ethereal coming back to life in a space dragon's stomach I won't stop you but I will judge you.

* Stops ethereal from crystalizing while eaten

Co-authored-by: itseasytosee <55666666+itseasytosee@users.noreply.github.com>