[MIRROR] Fixes gunboots shooting the wearer (#5345)
* Fixes gunboots shooting the wearer (#58651)

For whatever reason, gunboots could immediately shoot the person wearing them when they fired. This obviously isn't ideal, so this fixes that. Once the bullet has ricocheted off something, of course, they're once again fair game to be hit. This also kills the projectile_shooter component because it was pointless
Fixes: #58242 (Bullets fired from gunboots hit the wearer)

As if the trail of people trying to kill you for wearing this isn't bad enough, you don't need the shoes themselves doing you in

* Fixes gunboots shooting the wearer

Co-authored-by: Ryll Ryll <3589655+Ryll-Ryll@users.noreply.github.com>