[MIRROR] Remove gender-specific preferences (#5375)
* Remove gender-specific preferences (#58785)

Makes hair (which already was, I think?), underwear, facial hair, and undershirts all available to all genders.

Random character generation is not changed. Eventually, I want to remake random characters to look more believable, but that is out of scope.

Gender-specific preferences are an outdated trope that make the code harder to work with.

This change is also important as a total makeover of the preferences menu and the preference internals are going underway. It is significantly easier to work with when a preference can simply be a list of things you can choose, rather than having to depend on another preference to decide what they can show. While I could make this change at that time, it is also important that the new preferences menu keep parity, and avoid bikeshed on a trifling detail.

* Remove gender-specific preferences

* AAA

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Co-authored-by: Gandalf <jzo123@hotmail.com>